moved messages to simple_pub_sub_message
