+++ simple-pub-sub-message/Cargo.toml
@@ -6,3 +6,5 @@ edition = "2021"
 [dependencies]
 thiserror = "2.0.12"
 anyhow = "1.0.97"
+tokio = { version = "1", features = ["full", "tracing"] }
+log = "0.4.26"
