+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod client;
-pub mod message;
 pub mod server;
 pub mod stream;
 pub mod topics;
+pub use simple_pub_sub_message::header::Header;
+pub use simple_pub_sub_message::message;
+pub use simple_pub_sub_message::pkt::PktType;
